# api/openapi.yaml
openapi: 3.0.3
info:
  title: ComicBagi
  description: PHP Symfony-based comic hosting catalog full-stack.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: /api
paths:
...
  /rest/languages:
    get:
      tags:
        - Language
      summary: List language.
      operationId: listLanguage
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Language list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Language'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Language
      summary: Add language.
      operationId: addLanguage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLanguage'
        required: true
      responses:
        '201':
          description: Language added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/languages/{lang}:
    get:
      tags:
        - Language
      summary: Get language.
      operationId: getLanguage
      parameters:
        - name: lang
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Language gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Language
      summary: Update language.
      operationId: updateLanguage
      parameters:
        - name: lang
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLanguage'
        required: true
      responses:
        '200':
          description: Language updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Language'
        '204':
          description: Language unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Language
      summary: Delete language.
      operationId: deleteLanguage
      parameters:
        - name: lang
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Language deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
...
  /rest/websites/{websiteHost}/item-languages:
    get:
      tags:
        - Website
      summary: List website item language.
      operationId: listWebsiteItemLanguage
      parameters:
        - name: websiteHost
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Website item language list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WebsiteItemLanguage'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Website
      summary: Add website item language.
      operationId: addWebsiteItemLanguage
      parameters:
        - name: websiteHost
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewWebsiteItemLanguage'
        required: true
      responses:
        '201':
          description: Website item language added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteItemLanguage'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/websites/{websiteHost}/item-languages/{lang}:
    get:
      tags:
        - Website
      summary: Get website item language.
      operationId: getWebsiteItemLanguage
      parameters:
        - name: websiteHost
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Website item language gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteItemLanguage'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Website
      summary: Update website item language.
      operationId: updateWebsiteItemLanguage
      parameters:
        - name: websiteHost
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetWebsiteItemLanguage'
        required: true
      responses:
        '200':
          description: Website item language updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteItemLanguage'
        '204':
          description: Website item language unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Website
      summary: Delete website item language.
      operationId: deleteWebsiteItemLanguage
      parameters:
        - name: websiteHost
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Website item language deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/links:
    get:
      tags:
        - Link
      summary: List link.
      operationId: listLink
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
        - name: websiteHost
          in: query
          schema:
            type: array
            items:
              type: string
        - name: relativeReference
          in: query
          schema:
            type: array
            items:
              type: string
        - name: href
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Link list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Link
      summary: Add link.
      operationId: addLink
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLink'
        required: true
      responses:
        '201':
          description: Link added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/links/{href}:
    get:
      tags:
        - Link
      summary: Get link.
      operationId: getLink
      parameters:
        - name: href
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Link gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Link
      summary: Update link.
      operationId: updateLink
      parameters:
        - name: href
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLink'
        required: true
      responses:
        '200':
          description: Link updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Link'
        '204':
          description: Link unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Link
      summary: Delete link.
      operationId: deleteLink
      parameters:
        - name: href
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Link deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/links/{linkHref}/item-languages:
    get:
      tags:
        - Link
      summary: List link item language.
      operationId: listLinkItemLanguage
      parameters:
        - name: linkHref
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: order
          in: query
          schema:
            type: string
        - name: orderBy
          in: query
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Link item language list.
          headers:
            X-Total-Count:
              schema:
                type: integer
            X-Pagination-Limit:
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LinkItemLanguage'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Link
      summary: Add link item language.
      operationId: addLinkItemLanguage
      parameters:
        - name: linkHref
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewLinkItemLanguage'
        required: true
      responses:
        '201':
          description: Link item language added.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkItemLanguage'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /rest/links/{linkHref}/item-languages/{lang}:
    get:
      tags:
        - Link
      summary: Get link item language.
      operationId: getLinkItemLanguage
      parameters:
        - name: linkHref
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Link item language gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkItemLanguage'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Link
      summary: Update link item language.
      operationId: updateLinkItemLanguage
      parameters:
        - name: linkHref
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetLinkItemLanguage'
        required: true
      responses:
        '200':
          description: Link item language updated.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkItemLanguage'
        '204':
          description: Link item language unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Link
      summary: Delete link item language.
      operationId: deleteLinkItemLanguage
      parameters:
        - name: linkHref
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Link item language deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
components:
  schemas:
...
    Language:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        lang:
          type: string
        name:
          type: string
      required:
        - createdAt
        - lang
        - name
    NewLanguage:
      type: object
      properties:
        lang:
          type: string
        name:
          type: string
      required: 
        - lang
        - name
    SetLanguage:
      type: object
      properties:
        lang:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
...
    WebsiteItemLanguage:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        languageLang:
          type: string
        machineTranslate:
          type: integer
          nullable: true
      required:
        - createdAt
        - languageLang
    NewWebsiteItemLanguage:
      type: object
      properties:
        languageLang:
          type: string
        machineTranslate:
          type: integer
          nullable: true
      required:
        - languageLang
    SetWebsiteItemLanguage:
      type: object
      properties:
        languageLang:
          type: string
          nullable: true
        machineTranslate:
          type: integer
          nullable: true
    Link:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        websiteHost:
          type: string
        websiteName:
          type: string
        relativeReference:
          type: string
          nullable: true
        itemLanguageCount:
          type: integer
      required:
        - createdAt
        - websiteHost
        - websiteName
        - itemLanguageCount
    NewLink:
      type: object
      properties:
        websiteHost:
          type: string
        relativeReference:
          type: string
          nullable: true
      required: 
        - websiteHost
    SetLink:
      type: object
      properties:
        websiteHost:
          type: string
          nullable: true
        relativeReference:
          type: string
          nullable: true
    LinkItemLanguage:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        languageLang:
          type: string
        machineTranslate:
          type: integer
          nullable: true
      required:
        - createdAt
        - languageLang
    NewLinkItemLanguage:
      type: object
      properties:
        languageLang:
          type: string
        machineTranslate:
          type: integer
          nullable: true
      required:
        - languageLang
    SetLinkItemLanguage:
      type: object
      properties:
        languageLang:
          type: string
          nullable: true
        machineTranslate:
          type: integer
          nullable: true
    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message
  responses:
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT